backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12

# Database
sqlalchemy>=2.0.0

# Validation
pydantic>=2.9.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# PDF extraction (replaces PyPDF2)
pdfplumber>=0.11.0

# Ollama HTTP client
httpx>=0.27.0

# Vector store — FAISS has native Python 3.14 wheels, no pydantic dependency
faiss-cpu>=1.13.0

